requirements.txt
scikit-learn
sklearn-preprocessing
pandas